+++ test/strongbox/toc_test.clj
@@ -90,9 +90,9 @@ SomeAddon.lua")
                      :dirsize 0
                      :label "Addon Name"
                      :description "Description of the addon here"
-                     :interface-version 80205
-                     :-toc/game-track :retail
-                     :supported-game-tracks [:retail]
+                     :interface-version-list [80205 11302]
+                     :-toc/game-track-list [:retail :classic]
+                     :supported-game-tracks [:classic :retail]
                      :installed-version "1.6.1"
                      :source "wowinterface"
                      :source-id 12345
@@ -108,8 +108,8 @@ SomeAddon.lua")
                      :dirname "EveryAddon"
                      :label "EveryAddon *"
                      :description nil
-                     :interface-version constants/default-interface-version
-                     :-toc/game-track :retail
+                     :interface-version-list [constants/default-interface-version]
+                     :-toc/game-track-list [:retail]
                      :supported-game-tracks [:retail]
                      :installed-version nil}
 
@@ -119,9 +119,9 @@ SomeAddon.lua")
 
                  ;; classic interface version gets a :classic game-track
                  [{:interface constants/default-interface-version-classic}
-                  (merge base-case {:interface-version constants/default-interface-version-classic
+                  (merge base-case {:interface-version-list [constants/default-interface-version-classic]
                                     :supported-game-tracks [:classic]
-                                    :-toc/game-track :classic})]
+                                    :-toc/game-track-list [:classic]})]
 
                  ;; addon is in development
                  [{:version "@project-version@"} (merge base-case
@@ -138,9 +138,9 @@ SomeAddon.lua")
           defaults {:dirname "dirname"
                     :description nil
                     :installed-version nil
-                    :interface-version constants/default-interface-version
+                    :interface-version-list [constants/default-interface-version]
                     :supported-game-tracks [:retail]
-                    :-toc/game-track :retail}
+                    :-toc/game-track-list [:retail]}
           cases [;; wowinterface
                  [{:x-wowi-id "123"} {:label "dirname *" :name "dirname"
                                       :source "wowinterface" :source-id 123
@@ -208,6 +208,27 @@ SomeAddon.lua")
               :let [expected (merge expected defaults)]]
         (is (= expected (toc/parse-addon-toc given addon-dir)))))))
 
+(deftest parse-addon-toc--duplicate-interface-versions-removed
+  (let [case {:interface 10000 :#interface 10000}
+        expected [10000]
+        use-defaults false]
+    (is (= expected (:interface-version-list (toc/-parse-addon-toc case use-defaults))))))
+
+(deftest parse-addon-toc--use-defaults
+  (testing "with-defaults true"
+    (let [case {:title nil
+                :interface nil}
+          expected {:label " *"
+                    :interface-version-list [constants/default-interface-version]}]
+      (is (= expected (select-keys (toc/-parse-addon-toc case true) [:label :interface-version-list])))))
+
+  (testing "with-defaults false"
+    (let [case {:title nil
+                :interface nil}
+          expected {:label nil
+                    :interface-version-list []}]
+      (is (= expected (select-keys (toc/-parse-addon-toc case false) [:label :interface-version-list]))))))
+
 (deftest rm-trailing-version
   (testing "parsing of 'Title' attribute in toc file"
     (let [cases [["Grid" "Grid"] ;; no trailing version? no problems
@@ -259,3 +280,52 @@ SomeAddon.lua")
     (let [fixture (helper/fixture-path "questie--invalid.toc")
           raw-data (toc/read-toc-file fixture)]
       (is (nil? (toc/parse-addon-toc raw-data))))))
+
+(deftest parse-addon-toc--multiple-interface-versions
+  (testing "multiple interface versions are supported"
+    (let [fixture {:title "foo"
+                   :label "Foo"
+                   :description "Foo Bar"
+                   :dirname "Baz"
+                   :interface "100206, 40400, 11502"
+                   :installed-version "1.2.3"
+                   :supported-game-tracks []}
+
+          expected {:description "Foo Bar",
+                    :dirname "Baz",
+                    :installed-version nil,
+                    :interface-version-list [100206 40400 11502],
+                    :label "foo",
+                    :name "foo",
+                    :supported-game-tracks [:retail :classic-cata :classic],
+                    :-toc/game-track-list [:retail :classic-cata :classic]}]
+      (is (= expected (toc/parse-addon-toc fixture))))))
+
+(deftest parse-interface-value
+  (testing "interface values can be parsed into a set of game tracks"
+    (let [cases [[nil []]
+
+                 ;; ---
+
+                 ["", []]
+                 ["asdf", []]
+                 ["1", [1]]
+                 ["1,2", [1,2]]
+                 ["1, 2", [1,2]]
+                 ["1,2, 3", [1,2,3]]
+                 ["1,2,3,", [1,2,3]]
+                 [",1,2,3,", [1,2,3]]
+                 ["100206, 40400, 11502", [100206, 40400, 11502]]
+
+                 ;; integers are used by tests but not encouraged.
+                 [0, [0]]
+                 [1, [1]]
+                 [100206, [100206]]
+
+                 ;; dupes
+                 ["1,1", [1]],
+                 ["1, 1, 1", [1]],
+                 ["1,1,2,2,3", [1,2,3]]]]
+
+      (doseq [[given expected] cases]
+        (is (= expected (toc/parse-interface-value given)))))))